txlint/pyproject.toml
[tool.poetry]
name = "txlint"
version = "0.1.0"
description = ""
authors = ["Jonghyun Park <parjong@gmail.com>"]

[tool.poetry.dependencies]
# TODO Use ^3.6.2
python = "^3.8"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
